+++ app/src/main/java/ayds/winchester/songinfo/moredetails/fulllogic/DataBase.kt
@@ -25,42 +25,3 @@ private const val createArtistInfoTableQuery: String =
 private const val DATABASE_NAME="dictionary.db"
 private const val DATABASE_VERSION= 1
 
-class DataBase(context: Context?) : SQLiteOpenHelper(context, DATABASE_NAME, null, DATABASE_VERSION) {
-    override fun onCreate(db: SQLiteDatabase) {
-             db.execSQL(createArtistInfoTableQuery)
-    }
-
-    override fun onUpgrade(db: SQLiteDatabase, oldVersion: Int, newVersion: Int) {}
-
-    fun saveArtist(artist: String?, info: String?) {
-        val values = ContentValues()
-        values.put(ARTIST_COLUMN, artist)
-        values.put(INFO_COLUMN, info)
-        values.put(SOURCE_COLUMN, 1)
-        writableDatabase?.insert(ARTISTS_TABLE, null, values)
-    }
-
-    fun getInfo(artist: String): String? {
-        val cursor = getArtistCursor(artist)
-        return mapArtistInfo(cursor)
-    }
-
-    private fun getArtistCursor(artist: String) =
-        readableDatabase.query(
-            ARTISTS_TABLE,
-            projection,
-            "$ARTIST_COLUMN = ?",
-            arrayOf(artist),
-            null,
-            null,
-            "artist DESC")
-
-    private fun mapArtistInfo(cursor: Cursor): String? =
-        with(cursor) {
-            if (moveToNext()) {
-                getString(cursor.getColumnIndexOrThrow(INFO_COLUMN))
-            } else {
-                null
-            }
-        }
-}
