+++ app/src/main/java/ayds/winchester/songinfo/moredetails/fulllogic/data/RepositoryImpl.kt
@@ -4,9 +4,28 @@ import ayds.winchester.songinfo.moredetails.fulllogic.domain.Repository
 import ayds.winchester.songinfo.moredetails.fulllogic.domain.entities.WikipediaArtist
 import java.io.IOException
 
+private const val ARTISTS_TABLE = "artists"
+private const val ID_COLUMN = "id"
+private const val ARTIST_COLUMN = "artist"
+private const val INFO_COLUMN = "info"
+private const val SOURCE_COLUMN = "source"
+private val projection = arrayOf(
+    ID_COLUMN,
+    ARTIST_COLUMN,
+    INFO_COLUMN,
+)
+private const val createArtistInfoTableQuery: String =
+    "create table $ARTISTS_TABLE (" +
+            "$ID_COLUMN INTEGER PRIMARY KEY AUTOINCREMENT," +
+            "$ARTIST_COLUMN string," +
+            "$INFO_COLUMN string," +
+            "$SOURCE_COLUMN string)"
+private const val DATABASE_NAME="dictionary.db"
+private const val DATABASE_VERSION= 1
+
 internal class RepositoryImpl : Repository {
 
-    private fun searchArtistInfo(artistName: String): ayds.winchester.songinfo.moredetails.fulllogic.WikipediaArtist? {
+    private fun searchArtistInfo(artistName: String): WikipediaArtist? {
         var wikipediaArtist = getArtistFromLocalStorage(artistName)
         when {
             wikipediaArtist != null ->  wikipediaArtist.markArtistAsLocal()
